refact: separate survey_item, user_response, enums
