+++ lib/_internal/utils/answer_response_util.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:prapare/models/fhir_questionnaire/survey/export.dart';
-import 'package:prapare/models/fhir_questionnaire/survey/item_type.dart';
+import 'package:prapare/models/fhir_questionnaire/survey/enums/item_type.dart';
 import 'package:validators/validators.dart' as validate;
 import 'package:validators/sanitizers.dart' as sanitize;
