+++ lib/_internal/utils/item_type_util.dart
@@ -1,5 +1,5 @@
 import 'package:fhir/r4.dart';
-import 'package:prapare/models/fhir_questionnaire/survey/item_type.dart';
+import 'package:prapare/models/fhir_questionnaire/survey/enums/item_type.dart';
 
 class ItemTypeUtil {
   String getCodeFromItemType(ItemType itemType) =>
