+++ lib/models/fhir_questionnaire/survey/export.dart
@@ -1,4 +1,4 @@
 export 'answer.dart';
 export 'survey.dart';
-export 'survey_item.dart';
-export 'user_response.dart';
+export 'survey_item/survey_item.dart';
+export 'user_response/user_response.dart';
