+++ lib/models/fhir_questionnaire/survey/survey_item/survey_item.dart
@@ -1,9 +1,9 @@
 import 'package:fhir/r4.dart';
 import 'package:prapare/_internal/utils/item_type_util.dart';
-import 'package:prapare/models/fhir_questionnaire/survey/item_type.dart';
+import 'package:prapare/models/fhir_questionnaire/survey/enums/item_type.dart';
 
-import 'answer.dart';
-import 'qformat.dart';
+import '../answer.dart';
+import '../enums/qformat.dart';
 
 part 'item_group.dart';
 part 'question.dart';
