+++ lib/models/fhir_questionnaire/survey/user_response/response_type.dart
@@ -3,14 +3,14 @@ abstract class ResponseType {
   dynamic value;
 }
 
-//   /// In some cases, boolean answers are accepted
+/// In some cases, boolean answers are accepted
 class ResponseBoolean extends ResponseType {
   ResponseBoolean(bool value) : super(value);
 }
 
-//   /// In other cases, decimals are accepted
-//   /// While FHIR accepts this resource as a decimal...
-//   /// Our current data elements only needs integers
+/// In other cases, decimals are accepted
+/// While FHIR accepts this resource as a decimal...
+/// Our current data elements only needs integers
 class ResponseDecimal extends ResponseType {
   ResponseDecimal(double value) : super(value);
 }
@@ -19,18 +19,18 @@ class ResponseInteger extends ResponseType {
   ResponseInteger(int value) : super(value);
 }
 
-//   /// for any generic case that needs a string (should be short)
+/// for any generic case that needs a string (should be short)
 class ResponseString extends ResponseType {
   ResponseString(String value) : super(value);
 }
 
-//   /// for any generic case that needs a string, but can be long
+/// for any generic case that needs a string, but can be long
 class ResponseText extends ResponseType {
   ResponseText(String value) : super(value);
 }
 
-//   /// Dates, DateTimes and Times respectively, they will be parsed and any part
-//   /// of the value that is not indicated will be ignored
+/// Dates, DateTimes and Times respectively, they will be parsed and any part
+/// of the value that is not indicated will be ignored
 class ResponseDate extends ResponseType {
   ResponseDate(DateTime value) : super(value);
 }
