+++ lib/models/fhir_questionnaire/survey/user_response/user_response.dart
@@ -1,6 +1,6 @@
 import 'package:fhir/r4.dart';
 import 'package:flutter/foundation.dart';
-import 'package:prapare/models/fhir_questionnaire/survey/item_type.dart';
+import 'package:prapare/models/fhir_questionnaire/survey/enums/item_type.dart';
 
 part 'answer_response.dart';
